fix: `/` 产 Float —— 语言说了四年,执行器从没做
`docs/semantics.md` 写着 `/` 对 Int/Int 返回 Float,还解释了后果
(整数中点必须写 `math.floor((lo + hi) / 2)`);
`examples/syntax/operators.lk` 断言 `15 / 4 > 3.7`;类型检查器
拒绝 `let c: Int = a / b`。只有执行器不认 —— 两条后端都做整数除法。
于是同一个表达式有三个答案:

    println(7 / 2);                        → 3.5   (常量折叠)
    let a = 7; let b = 2; println(a / b);  → 3     (运行时)
    let c: Int = a / b;                    → 类型错误

常量折叠还会**用值决定类型**:`20 / 4` 折成 Int,`7 / 2` 折成 Float。

改的是执行器,不是规则 —— 规则在三处白纸黑字,而且是有道理的那个。
VM 的 DivInt、通用 dynamic_div、AOT 的 I64/I64 下降、lkrt 的
dyn_div 全部改为浮点除法;除零随之按 IEEE 给 inf/NaN(`math.inf`
本来就是语言里的值,`%` 对 Int 除零仍 raise)。

这条改动掀出了三个连带的洞:

1. **编译器会整个略掉 `math.floor(x)`**,只要参数"看起来是整数"。
   在 `/` 产 Int 时成立,现在不成立:`math.floor(114 / 10)` 答
   `11.4`。是 bench 的校验和与 Lua 对不上抓住的 —— 一个优化算错了数。
2. **事实系统也认为 `/` 产 Int**,据此给商发类型化融合 opcode:
   `let mid = (lo + hi) / 2; mid * 2` 降成 `MulIntI`,运行时硬报
   "expected Int lhs, got Float"。
3. 内联器的 `inline_expr_is_supported` 漏了 `Expr::Cast`(另外两处
   遍历都认),于是含 `as` 的函数一律不能内联。

`math.floor(a / b)` 现在**是**整数除法唯一的写法,所以给它一条
`FloorDivInt`(向下取整,`-7 / 2` 是 `-4`;非 Int 操作数按 f64 除后
取整),VM 与 AOT 各一份实现,逐个符号组合对齐过。不要求操作数被
证明是 Int —— 那个前提正好让最需要它的调用(`subtotal` 来自 map)
用不上;opcode 自己有正确兜底。artifact 版本 13→14。

顺带:`Bytes` 渲染成 `Bytes([104,105])` 而不是 `<Bytes 2 bytes>`
——别的容器都渲染内容,而这条此前**没有任何测试**盯着,所以它活过
了统一渲染器那一轮。三宿主语料补了它,并加了一个直接断言输出的
测试(证明语料不是"三家一致地都失败")。

门禁:workspace 112 绿、AOT 55/55、examples 全通过、
bench 几何均值 0.986x(本轮基线 1.011x;校验和全部匹配)。

+++ core/src/vm/compiler/tests/call_intrinsics.rs
@@ -234,7 +234,10 @@ fn compiler_inlines_direct_function_with_while_early_return() {
             let lo = 0;
             let hi = limit - 1;
             while (lo <= hi) {
-                let mid = (lo + hi) / 2;
+                // `/` yields a Float, so an integer midpoint has to say so.
+                // (`math.floor` is the idiom; this crate's tests have no
+                // standard library, so the cast stands in for it.)
+                let mid = ((lo + hi) / 2) as Int;
                 let value = mid * 2;
                 if value == target {
                     return mid;
